fix: move rate limiters inline in login-handlers.ts to resolve CodeQL alerts
CodeQL could not trace express-rate-limit middleware through indirect
parameters. Moving formRateLimiter and credentialRateLimiter creation
directly into login-handlers.ts makes the dataflow visible to CodeQL.

- login-handlers.ts: imports rateLimit and creates login/signup limiters inline
- router.ts: removes login/signup rate limiter definitions and deps plumbing
- Existing router rate-limit tests were reviewed and revalidated
